refactor(protocol): move body helpers into parent module
Move body filter helpers from a separate `body_helpers` submodule
into `mod.rs` so child modules access them from their parent rather
than importing from a sibling. Functions become private (no longer
`pub(super)`), matching Rust's visibility model.
